Edit: moved sphere scene to demos.
