+++ tp1/public/index.html
@@ -6,7 +6,7 @@
     body{ background-color: grey; }
     canvas{ background-color: white;}
   </style>
-  <script src="../dist/model/scene.js" type="module" defer></script>
+  <script src="../dist/demos/sphere-demo.js" type="module" defer></script>
   <meta name="viewport" content="width=device-width,initial-scale=1" />
   <meta name="description" content="" />
   <link rel="icon" href="assets/icon.ico">
